requirements.txt
crewai
crewai-tools
streamlit
python-dotenv
requests
folium
streamlit-folium